backend/bsw-api/requirements.txt
fastapi==0.110.1
uvicorn==0.23.2
gunicorn==20.1.0
httpx==0.27.2
langchain-openai==0.2.8
elasticsearch-dsl[async]==8.16.0
azure-storage-blob==12.24.0
langchain-text-splitters==0.3.2
Jinja2==3.1.4
SQLAlchemy==2.0.36
alembic==1.14.0
psycopg2==2.9.10
pytz===2024.2
pandas==2.2.2
stopwordsiso==0.6.1
nltk===3.9.1
langchain-mistralai==0.2.10
transformers==4.51.2
python-jose[cryptography]